set up database for reviews

+++ db/migrate/20220617032145_add_review_column.rb
@@ -1,5 +1,5 @@
 class AddReviewColumn < ActiveRecord::Migration[5.2]
   def change
-    add_column :products, :review, :integer, null:false
+    add_column :products, :review, :float, null:false
   end
 end
